refactor(react-native): use create-token auth endpoint instead of rn-pure-jwt to create token (#544)
Use the `https://stream-calls-dogfood.vercel.app/api/auth/create-token`
endpoint to create the token and populate it within the dogfooding app.
The usage of `react-native-pure-jwt` is removed as a part of the PR.
