+++ packages/react-native-dogfood/package.json
@@ -33,7 +33,6 @@
     "react-native-get-random-values": "^1.8.0",
     "react-native-incall-manager": "^4.0.0",
     "react-native-polyfill-globals": "^3.1.0",
-    "react-native-pure-jwt": "^3.0.1",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.1",
     "react-native-svg": "^13.6.0",
